chore(deps): update typescript resolution and checksum in yarn.lock (#1421)
* chore(deps): update typescript resolution and checksum in yarn.lock

Updated the resolution and checksum for typescript version 5.7.2 in yarn.lock to ensure consistency and integrity of the package.

* chore(deps): update @typescript-eslint packages to version 8.27.0

Updated the @typescript-eslint packages in yarn.lock to version 8.27.0, ensuring compatibility with TypeScript and improving overall package integrity.

* chore: update yarn command in GitHub Actions workflow to use --immutable flag

Modified the yarn command in the GitHub Actions workflow to include the --immutable flag, ensuring that the dependencies are not modified during the installation process.

* chore(deps): update typescript resolution hash and checksum in yarn.lock

Updated the resolution hash and checksum for typescript version 5.7.2 in yarn.lock to ensure package integrity and consistency.